requirements.txt
absl-py==2.1.0
accelerate==1.2.1
acvl-utils==0.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alembic==1.14.0
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
async-timeout==5.0.1
attrs==23.2.0
Automat==22.10.0
av==14.3.0
batchgenerators==0.25
bert-score==0.3.13
blinker==1.9.0
blis==0.7.11
brotlipy==0.7.0
buildtools==1.0.6
cachetools==5.3.2
catalogue==2.0.10
causal-conv1d==1.1.0
certifi @ file:///croot/certifi_1700501669400/work/certifi
cffi==1.17.1
chardet==5.2.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
cloudpathlib==0.20.0
cloudpickle==3.0.0
cmake==3.31.2
colorama==0.4.6
confection==0.1.5
connected-components-3d==3.19.0
constantly==23.10.4
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.10
Cython==3.0.11
datasets==3.5.0
deepspeed==0.13.5
Deprecated==1.2.14
dicom2nifti==2.4.10
dill==0.3.8
distro==1.9.0
docker==7.1.0
docker-pycreds==0.4.0
docopt==0.6.2
dotmap==1.3.30
einops==0.8.0
entrypoints==0.4
et-xmlfile==1.1.0
evaluate==0.4.3
exceptiongroup==1.2.2
fastremap==1.15.0
filelock==3.13.1
flash-attn==2.3.6
Flask==3.1.0
fonttools==4.51.0
frozenlist==1.5.0
fsspec==2023.12.2
ftfy==6.2.3
furl==2.1.3
future==1.0.0
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.42
google-auth==2.27.0
google-auth-oauthlib==1.2.0
graphene==3.4.3
graphql-core==3.2.5
graphql-relay==3.2.0
graphviz==0.20.3
greenlet==3.0.3
grpcio==1.60.1
gunicorn==22.0.0
h11==0.14.0
h5py==3.12.1
hjson==3.1.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.30.2
humanize==4.9.0
hyperlink==21.0.0
idna @ file:///croot/idna_1666125576474/work
imagecodecs==2024.9.22
imageio==2.34.0
importlib_metadata==7.2.1
incremental==22.10.0
iniconfig==2.0.0
iopath==0.1.10
itsdangerous==2.2.0
Jinja2==3.1.3
jiter==0.5.0
joblib==1.3.2
jsonpickle==4.0.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.5
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.3
linecache2==1.0.0
lit==18.1.8
lxml==5.3.0
Mako==1.3.8
marisa-trie==1.2.1
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
MedPy @ git+https://github.com/loli/medpy.git@54bb190e2a79141858df9d59be4496d2cba4bb96
medspacy==1.3.1
medspacy-quickumls==3.2
medspacy_unqlite==0.9.6
mkl-fft==1.3.1
mkl-random @ file:///home/builder/ci_310/mkl_random_1641843545607/work
mkl-service==2.4.0
mlflow==2.14.3
monai==1.3.0
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
mup==1.0.0
murmurhash==1.0.11
networkx==3.2.1
nibabel==5.2.0
nilearn==0.10.4
ninja==1.11.1.1
nltk==3.9.1
noise==1.2.2
nptyping==2.5.0
numpy==1.24.4
nvidia-ml-py==12.560.30
oauthlib==3.2.2
openai==1.45.0
opencv-python==4.9.0.80
openpyxl==3.1.2
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
orderedmultidict==1.0.1
packaging==24.1
pandas==2.0.3
peft==0.10.0
pillow==11.2.0
pluggy==1.5.0
portalocker==3.1.1
positional-encodings==6.0.3
preshed==3.0.9
propcache==0.3.1
protobuf==4.23.4
psutil==5.9.8
py-cpuinfo==9.0.0
pyarrow==15.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.9.1
pydantic_core==2.23.3
pydicom==2.4.4
PyFastNER==1.0.10
Pygments==2.17.2
pynrrd==1.0.0
pynvml==12.0.0
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pyparsing==3.1.2
PyRuSH==1.0.9
pysbd==0.3.4
pysimstring==1.2.1
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
pytest==8.3.4
python-dateutil==2.8.2
python-docx==1.1.2
python-gdcm==3.0.23.1
pytz==2024.1
PyYAML==6.0.2
querystring-parser==1.2.4
quicksectx==0.4.0
qwen-vl-utils==0.0.10
radgraph==0.1.17
RaTEScore==0.5.0
redo==2.0.4
referencing==0.35.1
regex==2023.12.25
requests==2.32.3
requests-oauthlib==1.3.1
rich==13.7.1
rouge_score==0.1.2
rpds-py==0.22.3
rsa==4.9
sacremoses==0.1.1
safetensors==0.5.1
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
sentence-transformers==2.5.1
sentencepiece==0.2.0
sentry-sdk==1.40.5
setproctitle==1.3.3
shellingham==1.5.4
SimpleITK==2.3.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
smart-open==7.1.0
smmap==5.0.1
sniffio==1.3.1
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.25
sqlparse==0.5.3
srsly==2.5.0
sympy==1.13.3
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
termcolor==2.5.0
thinc==8.2.5
threadpoolctl==3.3.0
tifffile==2024.2.12
tim==0.5
timm==1.0.9
tokenizers==0.21.1
tomli==2.2.1
torch==2.0.0
torchaudio==0.13.1
torchio==0.19.6
torchvision==0.15.1
tornado==6.4.1
tqdm==4.67.1
traceback2==1.4.0
transformers==4.49.0
triton==2.0.0
trl==0.16.1
Twisted==23.10.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2023.4
Unidecode==1.3.8
unittest2==1.1.0
urllib3 @ file:///croot/urllib3_1686163155763/work
vtk==9.3.0
wandb==0.16.3
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.16.0
xxhash==3.5.0
yacs==0.1.8
yarl==1.19.0
zipp==3.21.0
zope.interface==6.1